re-structure project folder
