--- .env.example
@@ -1,4 +1,0 @@
-SITE_URL=http://localhost:3000
-WOO_SITE_URL=http://yourwocommercesite.com
-WOO_CONSUMER_KEY=xxxxx
-WOO_SECRET=xxxxx
