+++ src/apollo/ApolloClient.js
@@ -4,7 +4,7 @@ import { InMemoryCache } from 'apollo-cache-inmemory';
 import { createHttpLink } from 'apollo-link-http';
 
 import { IntrospectionFragmentMatcher } from 'apollo-cache-inmemory';
-import introspectionQueryResultData from '../fragmentTypes';
+import introspectionQueryResultData from '../../fragmentTypes';
 
 // Fragment matcher.
 const fragmentMatcher = new IntrospectionFragmentMatcher({
