+++ src/components/checkout/CheckoutCartItem.js
