+++ src/queries/auth/register.js
@@ -1,0 +1,23 @@
+import gql from 'graphql-tag';
+/**
+ * Register user mutation query.
+ */
+export default gql`
+    mutation RegisterMyUser($username: String!, $email: String!, $password: String!) {
+        registerUser(
+            input: {
+                clientMutationId: "CreateUser"
+                username: $username
+                email: $email
+                password: $password
+            }
+        ) {
+            user {
+                id
+                name
+                email
+                nicename
+            }
+        }
+    }
+`;
