+++ src/queries/get-product-slug.js
@@ -1,0 +1,16 @@
+import gql from 'graphql-tag';
+
+export default gql`
+    query GET_PRODUCT_SLUGS {
+        products: products {
+            edges {
+                node {
+                    id
+                    databaseId
+                    name
+                    slug
+                }
+            }
+        }
+    }
+`;
