+++ src/queries/index.js
@@ -1,0 +1,5 @@
+export { default as PRODUCTS_QUERY } from './get-products';
+export { default as PRODUCT_QUERY } from './get-product';
+export { default as LOGIN_USER } from './auth/login';
+export { default as REGISTER_USER } from './auth/register';
+export { default as PRODUCT_SLUGS } from './get-product-slug';
