+++ src/styles/sass/layouts/_nprogress.scss
