+++ src/validator/isEmpty.js
