+++ src/validator/register.js
